backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Production RAG System Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "httpx>=0.26.0",
    # Vector store
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.0",
    # Local embeddings (works offline, no API key needed)
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    # Document processing
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "tiktoken>=0.5.0",
    # LLM clients
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    # Observability
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
